+++ CHANGELOG.md
@@ -7,7 +7,6 @@ Targets
 * Instructions for adding new functions and constraints in C.
 * Area/zone projections.
 * Variable generator curtailment cost.
-* Network object serialization using json.
 
 Unreleased
 ----------
@@ -31,6 +30,7 @@ Unreleased
 * Added "v_base" field of bus structure/object to store base voltage in kv and updated all parsers to store this info (done).
 * Made Network object pickle-able (done).
 * Added automatic enforcement of lower-triangularity of Hessian of objective functions (done).
+* Made Contingency object pickle-able (done).
 * Update Python wrapper documentation to show how to install with pip or download/run tests (todo).
 * Add cmake windows build script invokation in setup.py (todo).
 * Update examples, documentation and create release (todo).